Create how to use GitHub video
hi guys ! I found a good video on how to push/commit/make pull requests.

important to add changes to branches, once its finalized merge it to the main  branch.

-Creating a branch allows you to edit your project without changing the main branch. Now that you have a branch, it’s time to create a file and make your first commit!

What is a commit?: A commit is a set of changes to the files and folders in your project. A commit exists in a branch.

What is a pull request?: Collaboration happens on a pull request. The pull request shows the changes in your branch to other people and allows people to accept, reject, or suggest additional changes to your branch.

Pull requests let you tell others about changes you've pushed to a branch in a repository on GitHub. Once a pull request is opened, you can discuss and review the potential changes with collaborators and add follow-up commits before your changes are merged into the base branch.

You successfully created a pull request. You can now merge your pull request.

What is a merge?: A merge adds the changes in your pull request and branch into the main branch. For more information about merges, see "Merging a pull request."